requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
atlastk==0.13.5
certifi==2025.7.14
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
defusedxml==0.7.1
Django==5.2.6
django-dotenv==1.4.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
dotenv==0.9.9
drf-yasg==1.21.11
elementpath==5.0.3
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
flet==0.28.3
flet-web==0.28.3
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpie==3.2.4
httptools==0.6.4
httpx==0.28.1
idna==3.10
inflection==0.5.1
iso8601==2.1.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.6.4
oauthlib==3.3.1
packaging==25.0
psycopg2==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pypika-tortoise==0.6.1
PySocks==1.7.1
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
repath==0.9.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.4
sentry-sdk==2.33.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.47.2
tortoise==0.1.1
tortoise-orm==0.25.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
xmlschema==4.1.0
